+++ lib/genpass/genpass.h
@@ -74,6 +74,6 @@
 
 int genpass(uint8_t dk[64],
     const uint8_t * passwd, size_t passwdlen,
-    size_t maxmem, int megaops);
+    uint32_t maxmem, uint32_t megaops);
 
 #endif /* !_GENPASS_H_ */
